Improve S2437: message - replace silly with unnecessary
