+++ analyzers/tests/SonarAnalyzer.UnitTest/TestCases/UnnecessaryBitwiseOperation.CSharp9.cs
